Port PE file parsing to rust
Parse PE files in rust and expose methods to retrieve the PEInfo headers or relocated PE file contents. When a PE file payload is relocated, the original byte array containing the PE file contents is modified in-place. If no relocations are required, the payload remains unchanged. We no longer load the unmodified PE file into memory beofre patching it. Instead the PE file is stored on the Context and is modified there to avoid making all of the PE file code paths unsafe.

I have updated the simpleguest.exe test binary to force the compiled binary to contain a relocation to help test out my changes (I aliased println to foo and use the alias once).

The PE file tests have been updated to test a few more scenarios now that one of them contains a relocation now. This means the tests have some hard coded values based on the contents of the checked-in testdata binaries that would need to be changed if the files were recompiled. For example the entrypoint would probably change and maybe a few other addresses that we use to validate that the PE file was read correctly and relocated. So keep that in mind if you update the testdata binaries.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>