[hyperlight_guest] Ensure downstream crates can find libc headers
Previously, when building libc, hyperlight_guest didn't do anything to
preserve the relevant header files for downstream crates that might
need to build native library dependencies. This changes the build
script to copy the header files into $OUT_DIR/include, provide this
path in cargo metadata that downstream crates can consume (adding a
links key to the Cargo manifest in order to make this possible), and
builds clang and llvm-ml wrappers for the convenience of such
downstream crates.