[actions/CreateRelease] replace erroneous bash with PowerShell
When deciding whether or not to create a nuget release, we want to use
the same logic as when deciding whether or not to create a Cargo
release.  Unfortunately, the nuget release job runs on Windows, so we
can't use the same bash script that we use in the cargo release
one---we have to port it to pwsh.exe instead.